+++ sample-code/src/index.ts
@@ -2,7 +2,8 @@
 export {
   chatCompletion,
   computeEmbedding,
-  chatCompletionWithDestination
+  chatCompletionWithDestination,
+  chatCompletionWithFunctionCall
   // eslint-disable-next-line import/no-internal-modules
 } from './foundation-models/azure-openai.js';
 export {
@@ -28,7 +29,10 @@ export {
   invokeChain,
   invokeRagChain
 } from './langchain-azure-openai.js';
-export { invokeChain as orchestrationInvokeChain, invokeLangGraphChain } from './langchain-orchestration.js';
+export {
+  invokeChain as orchestrationInvokeChain,
+  invokeLangGraphChain
+} from './langchain-orchestration.js';
 export {
   getDeployments,
   getDeploymentsWithDestination,