+++ packages/document-grounding/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ai-sdk/document-grounding
 
+## 1.12.0
+
+### Patch Changes
+
+- @sap-ai-sdk/core@1.12.0
+
+## 1.11.0
+
+### Patch Changes
+
+- Updated dependencies [627a152]
+  - @sap-ai-sdk/core@1.11.0
+
 ## 1.10.0
 
 ### Patch Changes