+++ tests/e2e-tests/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ai-sdk/e2e-tests
 
+## 1.12.0
+
+### Patch Changes
+
+- Updated dependencies [5360c8c]
+  - @sap-ai-sdk/orchestration@1.12.0
+  - @sap-ai-sdk/sample-code@1.12.0
+  - @sap-ai-sdk/ai-api@1.12.0
+
+## 1.11.0
+
+### Patch Changes
+
+- Updated dependencies [a8d05b0]
+- Updated dependencies [264c5b0]
+  - @sap-ai-sdk/orchestration@1.11.0
+  - @sap-ai-sdk/ai-api@1.11.0
+  - @sap-ai-sdk/sample-code@1.11.0
+
 ## 1.10.0
 
 ### Patch Changes